+++ src/main/java/org/vshmaliukh/client_server_model/ConstantsForClientServerModel.java
@@ -1,0 +1,12 @@
+package org.vshmaliukh.client_server_model;
+
+public final class ConstantsForClientServerModel {
+
+    private ConstantsForClientServerModel(){}
+
+    public static final int SOCKET_PORT_NUMBER = 8888;
+    public static final int MAX_CONNECTION_NUMBER = 10;
+
+    public static final String HOST_NAME = "localhost";
+
+}
